+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 9.3
 - SQL use CONCAT() instead of pipes || for MySQL compatibility, program wide
 - Fix first item in the list not displayed in Accounting/includes/DailyTransactions.php
 - SQL time interval for MySQL compatibility in PasswordReset.php & index.php
+- SQL use CAST(X AS varchar(X)) instead of to_char() for MySQL compatibility in Dashboard.inc.php & Reminders.php
 
 Changes in 9.2.2
 ----------------